// src/OpenAPISDKGenerators.Frontend/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseURL: 'http://localhost:5000/' | (string & {});
};

export type Product = {
    id: string;
    title: string;
    description: string;
    price: number;
};

export type User = ({
    type?: 'student';
} & UserStudent) | ({
    type?: 'teacher';
} & UserTeacher);

export type UserStudent = {
    type?: 'student';
    gpa: number;
    major: string;
    id: string;
    name: string;
    email: string;
};

export type UserTeacher = {
    type?: 'teacher';
    department: string;
    yearsOfExperience: number;
    id: string;
    name: string;
    email: string;
};

export type GetProductsData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/products';
};

export type GetProductsResponses = {
    /**
     * OK
     */
    200: Array<Product>;
};

export type GetProductsResponse = GetProductsResponses[keyof GetProductsResponses];

export type CreateProductData = {
    body: Product;
    path?: never;
    query?: never;
    url: '/products';
};

export type CreateProductResponses = {
    /**
     * OK
     */
    200: Product;
};

export type CreateProductResponse = CreateProductResponses[keyof CreateProductResponses];

export type GetUsersData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/users';
};

export type GetUsersResponses = {
    /**
     * OK
     */
    200: Array<User>;
};

export type GetUsersResponse = GetUsersResponses[keyof GetUsersResponses];